requirements.txt
absl-py==0.15.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.1
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
clip==1.0
cloudpickle==2.0.0
colorama==0.4.4
cryptography==3.4.8
cssselect2==0.5.0
cycler==0.11.0
cytoolz==0.11.0
dask==2022.2.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
einops==0.3.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.1
ffmpeg-python==0.2.0
filelock==3.6.0
Fiona==1.8.22
Flask==1.1.2
frozenlist==1.2.0
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
geopandas==0.12.1
gitdb==4.0.7
GitPython==3.1.18
google-auth==2.6.0
google-auth-oauthlib==0.4.1
grpcio==1.42.0
h5py==2.10.0
huggingface-hub==0.5.1
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
importlib-metadata==4.11.3
importlib-resources==5.2.0
intel-openmp==2021.4.0
ipykernel==6.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyter-server-mathjax==0.2.3
jupyterlab==3.1.12
jupyterlab-git==0.32.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
kiwisolver==1.3.2
livelossplot==0.1.2
llvmlite==0.38.0
locket==0.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
mat73==0.46
matplotlib==3.3.1
matplotlib-inline==0.1.3
mhcnames==0.4.8
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
munch==2.5.0
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbdime==3.1.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.7.1
nomen==0.42
notebook==6.1.5
numba==0.55.1
numpy==1.19.2
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.5.5.64
openpyxl==3.0.10
openslide-python==1.1.2
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pip==21.2.2
prettytable==2.0.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
Pygments==2.11.2
PyJWT==2.1.0
pynndescent==0.5.4
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyproj==3.4.0
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pywin32==302
pywinpty==2.0.2
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.49
scikit-image==0.18.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.6.0
setproctitle==1.2.3
setuptools==61.2.0
Shapely==1.8.2
shortuuid==1.0.9
sip==4.19.13
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.6.0
tensorboardX==2.1
terminado==0.13.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.7.2
timm==0.5.4
tinycss2==1.1.1
tokenizers==0.12.1
toolz==0.11.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchfile==0.1.0
torchsummary==1.5.1
torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.19.0.dev0
typing_extensions==4.1.1
umap-learn==0.5.2
urllib3==1.26.9
visdom==0.1.8.9
vit-pytorch==0.20.8
wandb==0.12.20
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.3
wget==3.2
wheel==0.37.1
widgetsnbextension==3.5.2
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.3
zipp==3.7.0
